Fixes Chapter 4 issues and test in chapter 9
