--- Chapter04/Exercise 4.05/src/setupTests.js
@@ -1,5 +1,0 @@
-import { configure } from 'enzyme';
-import Adapter from 'enzyme-adapter-react-16';
-
-configure({ adapter: new Adapter() });
-
